Enable strict compilation warning flags
Warnings  incurred by these flags are also fixed.

Also fix the issue that --build-test doesn't work.

+++ test/tests/lru_cache_unittest.cpp
@@ -94,7 +94,7 @@ TEST_CASE("Put, get and eviction", "[LRUCache]")
 
         // case: LRU replacement when running out of free space
 
-        for (int i = 3; i < array_size(candidates); ++i) {
+        for (size_t i = 3; i < array_size(candidates); ++i) {
             alphabet.Put(candidates[i].first, candidates[i].second);
         }
